docs: Add comment explaining path resolution in book wrapper script
The bin/book wrapper script tries to resolve relative paths
from the current directory first. If that fails (path doesn't
exist relative to cwd), it passes the path as-is to PHP which
will try to resolve it relative to BOOK_STORAGE_PATH.

This allows:
- book . (current directory)
- book /absolute/path
- book relative/from/cwd
- book Action/Author/Book (relative to book root)

The PHP code now handles newlines in paths (in case they
somehow get introduced) by splitting on \r\n and processing
each part separately.